Fix getAvailQuantity of assets
When asking for the available quantity of a stock owned by a customer, if the
stock is not owned, return 0 instead of throwing an exception.

This matches the proper semantics, and allows us to check for 0 quantity
owned when making a sell trade, thereby allowing us to detect and
prevent short selling.